commit 0a9a25a293d437b1563e1d8479fef8f3795ba817
    13021 Invalid state if bindtextdomain() fails during re-binding
commit 5fae793b9b02afd1f3f434f3a915a64c08edc7b2
    13062 loader: make env_discard() public
commit b89fb824168e36795c325b5e0a4d1e71dd2d8d65
    13068 gptzfsboot: command from /boot/config should be nul terminated
commit 8548ec78d85644539a23c4262ed2b9512a47865e
    13065 log when suspending a zpool
commit 84971882a96ac0fecd538b02208054a872ff8af3
    12996 bhyve kernel should be wscheck clean
commit e0c0d44e917080841514d0dd031a696c74e8c435
    12989 improve interface boundary for bhyve MMIO
    12990 improve interface boundary for bhyve ins/outs
    12991 bhyve vlapic should SIPI more carefully
commit 76f19f5fdc974fe5be5c82a556e43a4df93f1de1
    12983 Want additional prototypes for manual pages
commit 341c5f490806c8b3e6e31512923db4c0e1b464b1
    13041 i40e_get_available_resources() broken again for X722 part
commit dcbbe9e03d9d35c714d909a1f4767ce83c885e80
    13049 cstyle should tolerate freebsd headers

+++ usr/src/uts/i86pc/io/vmm/io/vatpic.c
@@ -709,8 +709,8 @@ vatpic_write(struct vatpic *vatpic, struct atpic *atpic, bool in, int port,
 }
 
 int
-vatpic_master_handler(struct vm *vm, int vcpuid, bool in, int port, int bytes,
-    uint32_t *eax)
+vatpic_master_handler(struct vm *vm, int vcpuid, bool in, uint16_t port,
+    uint8_t bytes, uint32_t *eax)
 {
 	struct vatpic *vatpic;
 	struct atpic *atpic;
@@ -720,17 +720,17 @@ vatpic_master_handler(struct vm *vm, int vcpuid, bool in, int port, int bytes,
 
 	if (bytes != 1)
 		return (-1);
- 
+
 	if (in) {
 		return (vatpic_read(vatpic, atpic, in, port, bytes, eax));
 	}
- 
+
 	return (vatpic_write(vatpic, atpic, in, port, bytes, eax));
 }
 
 int
-vatpic_slave_handler(struct vm *vm, int vcpuid, bool in, int port, int bytes,
-    uint32_t *eax)
+vatpic_slave_handler(struct vm *vm, int vcpuid, bool in, uint16_t port,
+    uint8_t bytes, uint32_t *eax)
 {
 	struct vatpic *vatpic;
 	struct atpic *atpic;
@@ -749,8 +749,8 @@ vatpic_slave_handler(struct vm *vm, int vcpuid, bool in, int port, int bytes,
 }
 
 int
-vatpic_elc_handler(struct vm *vm, int vcpuid, bool in, int port, int bytes,
-    uint32_t *eax)
+vatpic_elc_handler(struct vm *vm, int vcpuid, bool in, uint16_t port,
+    uint8_t bytes, uint32_t *eax)
 {
 	struct vatpic *vatpic;
 	bool is_master;
